backend/app/Systéme_orientation_académique/requirements.txt
# --- API Server ---
fastapi
uvicorn[standard]
python-multipart

# --- Machine Learning & Data ---
scikit-learn
pandas
PyYAML
joblib